+++ lib/overcommit/hook/pre_commit/rubo_cop.rb
@@ -3,20 +3,31 @@ module Overcommit::Hook::PreCommit
   #
   # @see http://batsov.com/rubocop/
   class RuboCop < Base
-    MESSAGE_TYPE_CATEGORIZER = lambda do |type|
+    GENERIC_MESSAGE_TYPE_CATEGORIZER = lambda do |type|
+      type =~ /^warn/ ? :warning : :error
+    end
+
+    COP_MESSAGE_TYPE_CATEGORIZER = lambda do |type|
       type.include?('W') ? :warning : :error
     end
 
     def run
       result = execute(command, args: applicable_files)
       return :pass if result.success?
-      return [:fail, result.stderr] unless result.stderr.empty?
 
-      extract_messages(
+      generic_messages = extract_messages(
+        result.stderr.split("\n"),
+        /^(?<type>[a-z]+)/i,
+        GENERIC_MESSAGE_TYPE_CATEGORIZER,
+      )
+
+      cop_messages = extract_messages(
         result.stdout.split("\n"),
         /^(?<file>(?:\w:)?[^:]+):(?<line>\d+):[^ ]+ (?<type>[^ ]+)/,
-        MESSAGE_TYPE_CATEGORIZER,
+        COP_MESSAGE_TYPE_CATEGORIZER,
       )
+
+      generic_messages + cop_messages
     end
   end
 end